requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

disnake==2.5.2
gspread==5.7.2
oauth2client==4.1.3
python-dotenv==0.21.1
pytz==2022.7.1
